requirements.txt
# Core
Django==5.2.10
djangorestframework==3.16.1

# Auth
djangorestframework-simplejwt==5.5.0
djangorestframework-simplejwt[blacklist]==5.5.0

# API / docs
drf-spectacular==0.29.0

# WebSockets / ASGI
channels==4.3.2
daphne==4.2.1
channels-redis==4.3.0

# Async
celery==5.6.2
redis==7.1.0
django-redis==6.0.0

# Security / misc
django-cors-headers==4.9.0
django-csp==4.0
cryptography==46.0.4
Pillow==12.1.0
requests==2.32.5

# DB
psycopg2==2.9.11

# Resemble SDK
resemble==1.7.0
